requirements.txt
torch
torchvision
gsplat
numpy
Pillow
lpips
torchmetrics
PyYAML
omegaconf
tqdm
ninja
jaxtyping
timm
einops
gdown
